organize gpu package
