+++ lib/src/gpu/background/scene_background_builder.dart
@@ -1,9 +1,9 @@
 import 'package:flutter_scene/scene.dart';
 import 'package:vector_math/vector_math.dart';
-import 'package:vector_tile_renderer/src/gpu/background_geometry.dart';
+import 'package:vector_tile_renderer/src/gpu/background/background_geometry.dart';
 import 'package:vector_tile_renderer/src/gpu/colored_material.dart';
 
-import '../../vector_tile_renderer.dart';
+import '../../../vector_tile_renderer.dart';
 
 class SceneBackgroundBuilder {
   final Scene scene;
