+++ lib/src/gpu/line/line_geometry.dart
